HsInt=Int64, so use Int64
Our handling of ints (Int vs Int64 vs ...) is not very consistent.
We should handle them consistently as Int64, even if it is slightly more
painful with the casts to pass them to the appropriate functions.

+++ haskell-ffi/Sources/HaskellFFIMacros/ForeignImportHaskellMacro.swift
@@ -41,7 +41,7 @@ public struct ForeignImportHaskellMacro: PeerMacro {
                 else {
                     return """
                     var \(arg)_data = try hs_enc.encode(\(arg))
-                    let \(arg)_datalen = Int(\(arg)_data.count)
+                    let \(arg)_datalen = Int64(\(arg)_data.count)
                     """
                 }
             }).joined(separator: "\n"),
